MEGA commit, cause I'm lazy

+++ src/rules/cost.rs
@@ -5,11 +5,11 @@ pub(crate) struct Cost {
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) struct ManaCost {
-    pub(crate) generic: u32,
-    pub(crate) red: u32,
-    pub(crate) green: u32,
-    pub(crate) blue: u32,
-    pub(crate) white: u32,
-    pub(crate) black: u32,
-    pub(crate) colorless: u32,
+    pub(crate) generic: u8,
+    pub(crate) red: u8,
+    pub(crate) green: u8,
+    pub(crate) blue: u8,
+    pub(crate) white: u8,
+    pub(crate) black: u8,
+    pub(crate) colorless: u8,
 }
